backend/requirements.txt
# Flask and extensions
Flask==2.3.2
Flask-Cors==4.0.0
Werkzeug==2.3.6

# Database
pymongo==4.5.0

# Authentication
PyJWT==2.8.0

# Google Drive API
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# Environment variables
python-dotenv==1.0.0

# Utility packages
uuid==1.30
requests==2.31.0

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.0

# Logging
structlog==23.1.0